+++ language-support/scala/examples/quickstart-scala/application/src/main/scala/com/digitalasset/quickstart/iou/FutureUtil.scala
@@ -1,0 +1,11 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.digitalasset.quickstart.iou
+import scala.concurrent.Future
+
+object FutureUtil {
+  def toFuture[A](o: Option[A]): Future[A] =
+    o.fold(Future.failed[A](new IllegalStateException(s"Empty option: $o")))(a =>
+      Future.successful(a))
+}